Added cdp-table-conversion job and broke common functions into cdp-conversion-library
